Use loop to compute reciprocal
The Python prototype code to calculate the reciprocal of the
binary32/64/128 values now uses a loop. Inside the loop is the
code to perform one iteration of Newton's Method. The number of
times the loop is performed is controlled by the number of bits
in the significand.

This set us up to be able to use similar logic in the Verilog
module.